resume_analyzer/requirements.txt
# AI Resume Analyzer Dependencies
gradio>=4.0.0
openai>=1.0.0
pdfplumber>=0.10.0
python-docx>=1.0.0
pydantic>=2.0.0
tiktoken>=0.5.0
python-dotenv>=1.0.0
